code/models/requirements.txt
Pillow>=8.0.0
numpy>=1.19.0
torch>=1.9.0
torchvision>=0.10.0